fix: remove orphaned test files and rescue unique tests into internal module
Two test files under delivery/work_pulling_producer_controller/ and
delivery/producer_controller/ were never compiled because their parent
modules are pure type aliases with no `mod tests` declaration.

- Rescue 4 unique tests from the orphaned wppc file into
  internal/work_pulling_producer_controller/tests.rs
- Strengthen assert in durable_queue_store_ack test to verify
  ProducerControllerCommandKind::Msg variant
- Delete both orphaned test files and their empty directories
- producer_controller orphan was identical to internal; safe delete
